% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-methods.R
\name{add_question}
\alias{add_question}
\title{Add Question to Template}
\usage{
add_question(tmpl, q)
}
\arguments{
\item{tmpl}{template object}

\item{q}{question object}
}
\value{
The \code{markermd_template} object with \code{q} appended to its questions.
}
\description{
Add Question to Template
}
