% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-methods.R
\name{as.list.markermd_metadata}
\alias{as.list.markermd_metadata}
\title{Convert template_metadata to List}
\usage{
\method{as.list}{markermd_metadata}(x, ...)
}
\arguments{
\item{x}{template_metadata object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Convert template_metadata to List
}
