% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-methods.R
\name{as.list.markermd_node_selection}
\alias{as.list.markermd_node_selection}
\title{Convert node_selection to List}
\usage{
\method{as.list}{markermd_node_selection}(x, ...)
}
\arguments{
\item{x}{node_selection object}
}
\description{
Convert node_selection to List
}
