% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-methods.R
\name{as.list.markermd_question}
\alias{as.list.markermd_question}
\title{Convert question to List}
\usage{
\method{as.list}{markermd_question}(x, ...)
}
\arguments{
\item{x}{question object}
}
\description{
Convert question to List
}
