% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-methods.R
\name{as.list.markermd_template}
\alias{as.list.markermd_template}
\title{Convert template to List}
\usage{
\method{as.list}{markermd_template}(x, ...)
}
\arguments{
\item{x}{template object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Convert template to List
}
