% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-methods.R
\name{as_markermd_metadata}
\alias{as_markermd_metadata}
\title{Convert List to template_metadata}
\usage{
as_markermd_metadata(x)
}
\arguments{
\item{x}{List with metadata fields}
}
\description{
Convert List to template_metadata
}
