% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-methods.R
\name{as_markermd_node_selection}
\alias{as_markermd_node_selection}
\title{Convert List to node_selection}
\usage{
as_markermd_node_selection(x)
}
\arguments{
\item{x}{Integer vector or list with indices}
}
\description{
Convert List to node_selection
}
