% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-methods.R
\name{as_markermd_question}
\alias{as_markermd_question}
\title{Convert List to question}
\usage{
as_markermd_question(x)
}
\arguments{
\item{x}{List with id, name, selected_nodes fields}
}
\description{
Convert List to question
}
