% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ast_selectable.R
\name{ast_selectable_server}
\alias{ast_selectable_server}
\title{Selectable AST Server}
\usage{
ast_selectable_server(id, ast, selected_nodes = shiny::reactive(integer(0)))
}
\arguments{
\item{id}{Character. Module namespace ID}

\item{ast}{Reactive. The parsed AST object}

\item{selected_nodes}{Reactive. Currently selected node indices}
}
\description{
Selectable AST Server
}
