% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ast.R
\name{ast_selectable_ui}
\alias{ast_selectable_ui}
\title{Selectable AST UI}
\usage{
ast_selectable_ui(id, title = "Document Structure")
}
\arguments{
\item{id}{Character. Module namespace ID}

\item{title}{Character. Panel title (default: "Document Structure")}
}
\description{
Selectable AST UI
}
