% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tree.R
\name{build_ast_tree_structure}
\alias{build_ast_tree_structure}
\title{Tree Display Utilities}
\usage{
build_ast_tree_structure(ast)
}
\arguments{
\item{ast}{parsermd AST object}
}
\value{
List with tree structure information
}
\description{
Functions for creating hierarchical tree displays of AST structures
Build Tree Structure from AST
}
\details{
Create a hierarchical tree structure similar to parsermd's print output
}
