% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tree.R
\name{build_unified_tree_level}
\alias{build_unified_tree_level}
\title{Build Unified Tree Level}
\usage{
build_unified_tree_level(
  tree_items,
  target_depth,
  parent_index,
  all_selected_nodes,
  ns,
  directly_selected_nodes = integer(0),
  selection_mode = "readonly",
  id_prefix = NULL
)
}
\arguments{
\item{tree_items}{All tree items}

\item{target_depth}{Current depth level to build}

\item{parent_index}{Parent node index (NULL for root)}

\item{all_selected_nodes}{Vector of all selected node indices}

\item{ns}{Shiny namespace function}

\item{directly_selected_nodes}{Vector of directly selected node indices}

\item{selection_mode}{Character. Selection mode}

\item{id_prefix}{Character. Optional prefix for button IDs}
}
\value{
List of HTML elements for this level
}
\description{
Build a nested tree structure supporting different selection modes
}
