% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tree.R
\name{compute_all_selected_nodes}
\alias{compute_all_selected_nodes}
\title{Compute All Selected Nodes from Directly Selected}
\usage{
compute_all_selected_nodes(tree_items, directly_selected_nodes)
}
\arguments{
\item{tree_items}{List of tree items}

\item{directly_selected_nodes}{Vector of directly selected node indices}
}
\value{
Vector of all selected node indices (direct + descendants)
}
\description{
Given a list of directly selected nodes, compute the full list including all descendants
}
