% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parsing.R
\name{create_ast_tree_display}
\alias{create_ast_tree_display}
\title{Create AST Tree Display}
\usage{
create_ast_tree_display(ast)
}
\arguments{
\item{ast}{parsermd AST object}
}
\value{
Character vector with tree representation
}
\description{
Create a text representation of the AST similar to parsermd::print_tree()
}
