% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark-app.R
\name{create_cache_dir}
\alias{create_cache_dir}
\title{Create cache directory for artifacts}
\usage{
create_cache_dir(collection_path)
}
\arguments{
\item{collection_path}{Character string. Path to collection directory}
}
\value{
Character string. Path to cache directory
}
\description{
Create cache directory for artifacts
}
