% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mark_validate.R
\name{create_question_card}
\alias{create_question_card}
\title{Create Question Card}
\usage{
create_question_card(
  question,
  question_result,
  current_ast,
  original_ast,
  session
)
}
\arguments{
\item{question}{The question object}

\item{question_result}{The validation result for this question}

\item{current_ast}{The current document AST}

\item{original_ast}{The original template AST}

\item{session}{Shiny session object}
}
\value{
bslib card element
}
\description{
Creates a single question validation card
}
