% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rules.R
\name{create_rule_ui}
\alias{create_rule_ui}
\title{Create Rule UI}
\usage{
create_rule_ui(rule, ns, available_node_types, available_verbs)
}
\arguments{
\item{rule}{List containing rule data}

\item{ns}{Namespace function}

\item{available_node_types}{Vector of available node types}

\item{available_verbs}{Vector of available verbs}
}
\description{
Helper function to create UI for a single rule
}
