% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tree.R
\name{create_simple_tree}
\alias{create_simple_tree}
\title{Create Simple Tree View using ul/li elements}
\usage{
create_simple_tree(tree_items, selected_nodes, ns)
}
\arguments{
\item{tree_items}{List of tree items from build_ast_tree_structure}

\item{selected_nodes}{Vector of selected node indices}

\item{ns}{Shiny namespace function}
}
\value{
Shiny UI element with simple tree
}
\description{
Create a simple flat tree view for AST nodes
}
