% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rules.R
\name{create_verb_inputs_ui}
\alias{create_verb_inputs_ui}
\title{Create Verb Inputs UI}
\usage{
create_verb_inputs_ui(verb, rule_id, ns)
}
\arguments{
\item{verb}{Character. The selected verb}

\item{rule_id}{Numeric. The rule ID}

\item{ns}{Namespace function}
}
\description{
Helper function to create dynamic inputs based on selected verb
}
