% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_template.R
\name{evaluate_rule}
\alias{evaluate_rule}
\title{Evaluate Rule Against AST Subset}
\usage{
evaluate_rule(ast_subset, rule, num_selected_nodes = 0)
}
\arguments{
\item{ast_subset}{rmd_ast or list of nodes to evaluate}

\item{rule}{markermd_rule S7 object containing the rule definition}

\item{num_selected_nodes}{Integer number of originally selected heading nodes}
}
\value{
List with passed (logical) and message (character) components
}
\description{
Apply individual rule to an AST subset and determine if the rule condition is met.
}
