% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explore.R
\name{explore_server}
\alias{explore_server}
\title{Explore Server}
\usage{
explore_server(
  id,
  ast,
  current_repo_name = shiny::reactiveVal(NULL),
  validation_results = shiny::reactiveVal(NULL),
  selected_question_name = shiny::reactiveVal(NULL),
  template = shiny::reactiveVal(NULL)
)
}
\arguments{
\item{id}{Character. Module namespace ID}

\item{ast}{Reactive. The parsed AST object}

\item{current_repo_name}{Reactive. Current repository name}

\item{validation_results}{Reactive. Validation results for current repository}

\item{selected_question_name}{Reactive. Currently selected question name}

\item{template}{Reactive. Template object for reference}
}
\description{
Explore Server
}
