% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explore.R
\name{explore_ui}
\alias{explore_ui}
\title{Explore Interface Module}
\usage{
explore_ui(id)
}
\arguments{
\item{id}{Character. Module namespace ID}
}
\description{
Shiny module for exploring assignment documents
Explore UI
}
