% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tree.R
\name{find_all_descendants}
\alias{find_all_descendants}
\title{Find All Descendants of Node}
\usage{
find_all_descendants(tree_items, parent_node_index)
}
\arguments{
\item{tree_items}{List of tree items}

\item{parent_index}{Index of parent node}
}
\value{
Vector of all descendant node indices
}
\description{
Find all descendant nodes (children, grandchildren, etc.) for a given parent node
}
