% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tree.R
\name{find_node_children}
\alias{find_node_children}
\title{Find Children of Node}
\usage{
find_node_children(tree_items, parent_node_index)
}
\arguments{
\item{tree_items}{List of tree items}

\item{parent_index}{Index of parent node}
}
\value{
Vector of child node indices
}
\description{
Find all child nodes for a given parent node
}
