% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tree.R
\name{find_selected_descendants}
\alias{find_selected_descendants}
\title{Find Which Selected Nodes Are Descendants of a Given Node}
\usage{
find_selected_descendants(tree_items, new_node_index, current_selected)
}
\arguments{
\item{tree_items}{List of tree items}

\item{new_node_index}{Index of the newly selected node}

\item{current_selected}{Vector of currently selected node indices}
}
\value{
Vector of currently selected nodes that are descendants of the new node
}
\description{
Given a list of currently selected nodes and a new node, find which selected nodes
are descendants of the new node (and should be removed)
}
