% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_template.R
\name{generate_template_summary}
\alias{generate_template_summary}
\title{Generate Template Summary}
\usage{
generate_template_summary(questions, node_selections, ast_nodes = NULL)
}
\arguments{
\item{questions}{List of question objects}

\item{node_selections}{List mapping question IDs to selected nodes}

\item{ast_nodes}{List of AST nodes for context}
}
\value{
Character vector with template summary
}
\description{
Create a human-readable summary of the template
}
