% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-helpers.R
\name{get_allowed_filter_condition_types}
\alias{get_allowed_filter_condition_types}
\title{Get allowed filter condition types}
\usage{
get_allowed_filter_condition_types()
}
\value{
Character vector of allowed filter condition types
}
\description{
Returns the list of valid condition types that can be used in
question filters. Each type corresponds to a q2r predicate: "node type"
becomes an \verb{is(<pandoc class>)} test while the others map directly onto the
q2r mask helpers of the same name.
}
