% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule-helpers.R
\name{get_allowed_rule_verbs}
\alias{get_allowed_rule_verbs}
\title{Get allowed verbs for rules}
\usage{
get_allowed_rule_verbs()
}
\value{
Character vector of allowed rule verbs
}
\description{
Returns the list of valid verification verbs that can be used in validation rules.
Each verb corresponds to a different type of validation check.
}
