% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark-app.R
\name{get_archive_metadata}
\alias{get_archive_metadata}
\title{Get archive metadata from GitHub}
\usage{
get_archive_metadata(github_repos)
}
\arguments{
\item{github_repos}{Character vector. GitHub repositories in format "owner/repo"}
}
\value{
Data frame with repository and artifact metadata
}
\description{
Get archive metadata from GitHub
}
