% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{get_cached_artifact_path}
\alias{get_cached_artifact_path}
\title{Get cached artifact path}
\usage{
get_cached_artifact_path(collection_path, repo_name)
}
\arguments{
\item{collection_path}{Character string. Path to collection directory}

\item{repo_name}{Character string. Repository name}
}
\value{
Character string. Path to cached artifact file
}
\description{
Get cached artifact path
}
