% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-helpers.R
\name{get_default_filter_condition_value}
\alias{get_default_filter_condition_value}
\title{Get default value for a filter condition type}
\usage{
get_default_filter_condition_value(type)
}
\arguments{
\item{type}{Character. The condition type}
}
\value{
Default value appropriate for the condition type
}
\description{
Returns an appropriate default value for each filter condition
type. Used when creating new conditions or resetting after a type change.
}
