% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parsing.R
\name{get_document_summary}
\alias{get_document_summary}
\title{Get Document Summary}
\usage{
get_document_summary(ast)
}
\arguments{
\item{ast}{parsermd AST object}
}
\value{
List with document summary information
}
\description{
Extract basic information about the parsed document
}
