% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tree.R
\name{has_selected_ancestor}
\alias{has_selected_ancestor}
\title{Check if Node Has Selected Ancestor}
\usage{
has_selected_ancestor(tree_items, node_index, directly_selected_nodes)
}
\arguments{
\item{tree_items}{List of tree items}

\item{node_index}{Index of the node to check}

\item{directly_selected_nodes}{Vector of directly selected node indices}
}
\value{
Logical indicating if node has selected ancestors
}
\description{
Check if any ancestor node of the given node is in the directly selected list
}
