% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mark_rubric_item.R
\name{mark_rubric_item_server}
\alias{mark_rubric_item_server}
\title{Mark Rubric Item Server}
\usage{
mark_rubric_item_server(id, initial_item)
}
\arguments{
\item{id}{Character. Module namespace ID}

\item{initial_item}{markermd_rubric_item S7 object with initial state}
}
\description{
Server logic for rubric item with state management using S7 class
}
