% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mark_rubric_item.R
\name{mark_rubric_item_ui}
\alias{mark_rubric_item_ui}
\title{Mark Rubric Item Module}
\usage{
mark_rubric_item_ui(id, rubric_item)
}
\arguments{
\item{id}{Character. Module namespace ID}

\item{rubric_item}{markermd_rubric_item S7 object containing item properties}
}
\description{
Shiny module for individual rubric items with hotkey selection and point values
Mark Rubric Item UI
}
\details{
Creates UI for a single rubric item with hotkey button and description
}
