% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mark_rubric.R
\name{mark_rubric_server}
\alias{mark_rubric_server}
\title{Mark Rubric Server}
\usage{
mark_rubric_server(
  id,
  template,
  artifact_status_reactive,
  collection_path,
  use_qmd,
  collection,
  on_question_change = NULL
)
}
\arguments{
\item{id}{Character. Module namespace ID}

\item{template}{markermd_template. Static template object containing questions}

\item{artifact_status_reactive}{Reactive value. Artifact status for repositories}

\item{collection_path}{Character string. Path to collection directory}

\item{use_qmd}{Logical. Whether to parse .qmd files (TRUE) or .Rmd files (FALSE)}

\item{collection}{Parsed collection data from parsermd}

\item{on_question_change}{Reactive function. Callback when question selection changes}
}
\description{
Mark Rubric Server
}
