% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mark_rubric.R
\name{mark_rubric_ui}
\alias{mark_rubric_ui}
\title{Mark Rubric Interface Module}
\usage{
mark_rubric_ui(id)
}
\arguments{
\item{id}{Character. Module namespace ID}
}
\description{
Shiny module for displaying and navigating rubric questions during marking
Mark Rubric UI
}
