% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mark_validate.R
\name{mark_validate_ui}
\alias{mark_validate_ui}
\title{Mark Validation Interface Module}
\usage{
mark_validate_ui(id)
}
\arguments{
\item{id}{Character. Module namespace ID}
}
\description{
Shiny module for displaying validation results during marking
Mark Validation UI
}
