% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-classes.R
\name{markermd_template}
\alias{markermd_template}
\title{Markermd Template}
\usage{
markermd_template(
  original_ast = NULL,
  questions = list(),
  metadata = markermd_metadata()
)
}
\arguments{
\item{original_ast}{pandoc. The original parsed AST from q2r}

\item{questions}{List of question objects}

\item{metadata}{template_metadata. Template metadata}
}
\value{
A \code{markermd_template} S7 object.
}
\description{
S7 class representing a complete markermd template with questions and metadata
}
