% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_marking.R
\name{marking_ui}
\alias{marking_ui}
\title{Marking Interface Module}
\usage{
marking_ui(id)
}
\arguments{
\item{id}{Character. Module namespace ID}
}
\description{
Shiny module for grading assignments
Marking UI
}
