% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-classes.R
\name{new_markermd_filter_group}
\alias{new_markermd_filter_group}
\title{Create a new markermd filter group}
\usage{
new_markermd_filter_group(conditions = NULL, negate = FALSE)
}
\arguments{
\item{conditions}{List of markermd_filter_condition objects (if NULL, a
single default condition is used)}

\item{negate}{Logical. Whether the group is negated (defaults to FALSE)}
}
\value{
markermd_filter_group object
}
\description{
Convenience function to create a new filter group. By default
the group contains a single default condition.
}
