% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark-app.R
\name{open_folder}
\alias{open_folder}
\title{Open folder in system file manager (cross-platform)}
\usage{
open_folder(folder_path)
}
\arguments{
\item{folder_path}{Character string. Path to folder to open}
}
\description{
Open folder in system file manager (cross-platform)
}
