% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_question.R
\name{question_ui}
\alias{question_ui}
\title{Question Module}
\usage{
question_ui(id, name_id)
}
\arguments{
\item{id}{Character. Module namespace ID}

\item{name_id}{Integer. The question ID used for default naming}
}
\description{
Shiny module for managing a single question with its rules
Question UI
}
