% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rule.R
\name{rule_server}
\alias{rule_server}
\title{Rule Server}
\usage{
rule_server(id, initial_rule = NULL)
}
\arguments{
\item{id}{Character. Module namespace ID}

\item{initial_rule}{List. Initial rule data (optional)}
}
\description{
Rule Server
}
