% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rule.R
\name{rule_ui}
\alias{rule_ui}
\title{Rule Module}
\usage{
rule_ui(id, initial_rule = NULL)
}
\arguments{
\item{id}{Character. Module namespace ID}

\item{initial_rule}{markermd_rule object. Initial rule for setting default values (optional)}
}
\description{
Shiny module for managing a validation rule
Rule UI
}
