% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rules.R
\name{rules_server}
\alias{rules_server}
\title{Rules Server}
\usage{
rules_server(id, question_id, add_rule_trigger = NULL, initial_rules = NULL)
}
\arguments{
\item{id}{Character. Module namespace ID}

\item{question_id}{Reactive. The current question ID}

\item{add_rule_trigger}{Reactive. External trigger for adding rules}

\item{initial_rules}{List. Initial rules to load (optional)}
}
\description{
Rules Server
}
