% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rules.R
\name{rules_ui}
\alias{rules_ui}
\title{Rules Module}
\usage{
rules_ui(id)
}
\arguments{
\item{id}{Character. Module namespace ID}
}
\description{
Shiny module for creating and managing validation rules for questions
Rules UI
}
