% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_template.R
\name{save_template}
\alias{save_template}
\title{Save Template to File}
\usage{
save_template(questions, node_selections, file_path, metadata = list())
}
\arguments{
\item{questions}{List of question objects}

\item{node_selections}{List mapping question IDs to selected nodes}

\item{file_path}{Character. Path where to save the template}

\item{metadata}{List. Additional metadata about the template}
}
\value{
Logical indicating success
}
\description{
Save a template with questions and node mappings to a JSON file
}
