% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_git.R
\name{setup_assignment_repo}
\alias{setup_assignment_repo}
\title{Git Operations Utilities}
\usage{
setup_assignment_repo(
  assignment_path,
  local_dir = NULL,
  is_github_repo = FALSE
)
}
\description{
Functions for handling git repositories in markermd
}
