% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template_app}
\alias{template_app}
\title{Template App}
\usage{
template_app(ast, template_obj = NULL)
}
\arguments{
\item{ast}{Reactive. The parsed AST object}

\item{template_obj}{markermd_template S7 object. Optional template to load on startup}
}
\description{
Main Shiny application for creating assignment templates
Template App UI
}
