% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-app.R
\name{template_app_standalone}
\alias{template_app_standalone}
\title{Standalone Template App with Navbar}
\usage{
template_app_standalone(ast, template_obj = NULL, assignment_path = NULL)
}
\arguments{
\item{ast}{Reactive. The parsed AST object}

\item{template_obj}{markermd_template S7 object. Optional template to load on startup}

\item{assignment_path}{Character. Path to display in footer}
}
\description{
Template app wrapped in navbar for standalone use
}
