% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_template.R
\name{template_ui}
\alias{template_ui}
\title{Template Interface Module}
\usage{
template_ui(id)
}
\arguments{
\item{id}{Character. Module namespace ID}
}
\description{
Shiny module for creating assignment templates
Template UI
}
