% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_git.R
\name{validate_assignment_file}
\alias{validate_assignment_file}
\title{Validate Assignment File}
\usage{
validate_assignment_file(repo_path, filename)
}
\arguments{
\item{repo_path}{Character. Path to the assignment repository}

\item{filename}{Character. Name of the assignment file}
}
\value{
Character path to the assignment file
}
\description{
Check if the specified assignment file exists and is valid
}
