% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_template.R
\name{validate_template}
\alias{validate_template}
\title{Validate Template}
\usage{
validate_template(questions, node_selections, total_nodes)
}
\arguments{
\item{questions}{List of question objects}

\item{node_selections}{List mapping question IDs to selected nodes}

\item{total_nodes}{Integer. Total number of nodes in document}
}
\value{
List with validation results
}
\description{
Check if a template is valid and complete
}
