% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ast.R
\name{ast_readonly_ui}
\alias{ast_readonly_ui}
\title{Read-Only AST UI}
\usage{
ast_readonly_ui(id, title = "Document Structure")
}
\arguments{
\item{id}{Character. Module namespace ID}

\item{title}{Character. Panel title (default: "Document Structure")}
}
\description{
Read-Only AST UI
}
