% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{create_markermd_app}
\alias{create_markermd_app}
\title{Create Shiny App for markermd}
\usage{
create_markermd_app()
}
\value{
Shiny app object
}
\description{
Internal function to create the Shiny application
}
