% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_template.R
\name{create_question_templates}
\alias{create_question_templates}
\title{Create Question Templates}
\usage{
create_question_templates(template_obj)
}
\arguments{
\item{template_obj}{markermd_template S7 object}
}
\value{
Named list of parsermd templates
}
\description{
Creates parsermd templates from a template S7 object
}
