% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule-helpers.R
\name{get_allowed_node_types}
\alias{get_allowed_node_types}
\title{Get allowed node types for rules}
\usage{
get_allowed_node_types()
}
\value{
Character vector of allowed node types
}
\description{
Returns the list of valid node types that can be used in validation rules.
This includes both a catch-all "Any node" option and specific parsermd node types.
}
