% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-methods.R
\name{get_question}
\alias{get_question}
\title{Get Question by ID}
\usage{
get_question(tmpl, id)
}
\arguments{
\item{tmpl}{template object}

\item{id}{Integer question ID}
}
\description{
Get Question by ID
}
