% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark}
\alias{mark}
\title{Launch the markermd Shiny Application}
\usage{
mark(assignment_path, local_dir = NULL, filename, ...)
}
\arguments{
\item{assignment_path}{Character string. Path to local directory or GitHub repo in format "owner/repo"}

\item{local_dir}{Character string. Local directory for cloning (required for remote repos)}

\item{filename}{Character string. Name of the Rmd/qmd file to grade}

\item{...}{Additional arguments passed to shiny::runApp()}
}
\value{
Launches Shiny application
}
\description{
Launch the markermd Shiny Application
}
\examples{
\dontrun{
# Local assignment
mark("/path/to/assignment", filename = "homework.Rmd")

# Remote GitHub repo
mark("username/repo-name", local_dir = "/tmp/grading", filename = "assignment.qmd")
}
}
