% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-classes.R
\name{markermd_node_selection}
\alias{markermd_node_selection}
\title{Node Selection for Questions}
\usage{
markermd_node_selection(indices = integer(0))
}
\arguments{
\item{indices}{Integer vector of 1-based AST node indices}
}
\description{
S7 class representing selected AST nodes for a template question
}
