% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_marking.R
\name{marking_server}
\alias{marking_server}
\title{Marking Server}
\usage{
marking_server(id, ast, template)
}
\arguments{
\item{id}{Character. Module namespace ID}

\item{ast}{Reactive. The parsed AST object}

\item{template}{Reactive. The template object with questions}
}
\description{
Marking Server
}
