% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_marking.R
\name{marking_server}
\alias{marking_server}
\title{Marking Server}
\usage{
marking_server(
  id,
  ast,
  template_obj,
  validation_results = shiny::reactiveVal(NULL)
)
}
\arguments{
\item{id}{Character. Module namespace ID}

\item{ast}{Reactive. The parsed AST object}

\item{template_obj}{Reactive. markermd_template S7 object with questions and rules}

\item{validation_results}{Reactive. Optional validation results for current repository}
}
\description{
Marking Server
}
