% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parsing.R
\name{parse_assignment_document}
\alias{parse_assignment_document}
\title{Document Parsing Utilities}
\usage{
parse_assignment_document(file_path)
}
\description{
Functions for parsing Rmd/qmd documents using parsermd
}
