% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_question.R
\name{question_server}
\alias{question_server}
\title{Question Server}
\usage{
question_server(id, ast, initial_question = NULL)
}
\arguments{
\item{id}{Character. Module namespace ID}

\item{ast}{Reactive. The parsed AST object for building tree structure}

\item{question_id}{Integer. The question ID}
}
\description{
Question Server
}
