% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-methods.R
\name{remove_question}
\alias{remove_question}
\title{Remove Question from Template}
\usage{
remove_question(tmpl, id)
}
\arguments{
\item{tmpl}{template object}

\item{id}{Integer question ID to remove}
}
\description{
Remove Question from Template
}
