% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_template.R
\name{template_server}
\alias{template_server}
\title{Template Server}
\usage{
template_server(id, ast, template_path = NULL)
}
\arguments{
\item{id}{Character. Module namespace ID}

\item{ast}{Reactive. The parsed AST object}

\item{template_path}{Character. Optional path to template RDS file to load on startup}
}
\description{
Template Server
}
