% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_template.R
\name{template_server}
\alias{template_server}
\title{Template Server}
\usage{
template_server(id, ast)
}
\arguments{
\item{id}{Character. Module namespace ID}

\item{ast}{Reactive. The parsed AST object}
}
\description{
Template Server
}
