% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule-helpers.R
\name{validate_node_type}
\alias{validate_node_type}
\title{Validate node type value}
\usage{
validate_node_type(node_type)
}
\arguments{
\item{node_type}{Character. The node type to validate}
}
\value{
Character error message if invalid, NULL if valid
}
\description{
Checks if a node type is valid according to the allowed types.
}
